Wasm: support file IO and random generator in WASI

+++ core/src/main/java/org/teavm/backend/wasm/wasi/SizeResult.java
@@ -18,5 +18,5 @@ package org.teavm.backend.wasm.wasi;
 import org.teavm.interop.Structure;
 
 public class SizeResult extends Structure {
-    public int value;
+    public long value;
 }
